+++ packages/shared_preferences/shared_preferences_android/lib/src/shared_preferences_async_android.dart
@@ -18,10 +18,16 @@ base class SharedPreferencesAsyncAndroid
     extends SharedPreferencesAsyncPlatform {
   /// Creates a new plugin implementation instance.
   SharedPreferencesAsyncAndroid({
-    @visibleForTesting SharedPreferencesAsyncApi? api,
-  }) : _api = api ?? SharedPreferencesAsyncApi();
+    @visibleForTesting SharedPreferencesAsyncApi? dataStoreApi,
+    @visibleForTesting SharedPreferencesAsyncApi? sharedPreferencesApi,
+  })  : _dataStoreApi = dataStoreApi ??
+            SharedPreferencesAsyncApi(messageChannelSuffix: 'data_store'),
+        _sharedPreferencesApi = sharedPreferencesApi ??
+            SharedPreferencesAsyncApi(
+                messageChannelSuffix: 'shared_preferences');
 
-  final SharedPreferencesAsyncApi _api;
+  final SharedPreferencesAsyncApi _dataStoreApi;
+  final SharedPreferencesAsyncApi _sharedPreferencesApi;
 
   /// Registers this class as the default instance of [SharedPreferencesAsyncPlatform].
   static void registerWith() {
@@ -31,7 +37,19 @@ base class SharedPreferencesAsyncAndroid
   /// Returns a SharedPreferencesPigeonOptions for sending to platform.
   SharedPreferencesPigeonOptions _convertOptionsToPigeonOptions(
       SharedPreferencesOptions options) {
+    if (options is SharedPreferencesAsyncAndroidOptions) {
+      return SharedPreferencesPigeonOptions(
+        fileName: options.originalSharedPreferencesOptions?.fileName,
+        useDataStore:
+            options.backend == SharedPreferencesAndroidBackendLibrary.DataStore,
+      );
+    }
     return SharedPreferencesPigeonOptions();
+  }
+
+  SharedPreferencesAsyncApi _getApiForBackend(
+      SharedPreferencesPigeonOptions options) {
+    return options.useDataStore ? _dataStoreApi : _sharedPreferencesApi;
   }
 
   @override
@@ -40,13 +58,13 @@ base class SharedPreferencesAsyncAndroid
     SharedPreferencesOptions options,
   ) async {
     final PreferencesFilters filter = parameters.filter;
-    // TODO(tarrinneal): Remove cast once https://github.com/flutter/flutter/issues/97848
-    // is fixed. In practice, the values will never be null, and the native implementation assumes that.
-    return (await _api.getKeys(
+    final SharedPreferencesPigeonOptions pigeonOptions =
+        _convertOptionsToPigeonOptions(options);
+    final SharedPreferencesAsyncApi api = _getApiForBackend(pigeonOptions);
+    return (await api.getKeys(
       filter.allowList?.toList(),
-      _convertOptionsToPigeonOptions(options),
+      pigeonOptions,
     ))
-        .cast<String>()
         .toSet();
   }
 
@@ -60,8 +78,11 @@ base class SharedPreferencesAsyncAndroid
       throw ArgumentError(
           'StorageError: This string cannot be stored as it clashes with special identifier prefixes');
     }
+    final SharedPreferencesPigeonOptions pigeonOptions =
+        _convertOptionsToPigeonOptions(options);
+    final SharedPreferencesAsyncApi api = _getApiForBackend(pigeonOptions);
 
-    return _api.setString(key, value, _convertOptionsToPigeonOptions(options));
+    return api.setString(key, value, pigeonOptions);
   }
 
   @override
@@ -70,7 +91,10 @@ base class SharedPreferencesAsyncAndroid
     int value,
     SharedPreferencesOptions options,
   ) async {
-    return _api.setInt(key, value, _convertOptionsToPigeonOptions(options));
+    final SharedPreferencesPigeonOptions pigeonOptions =
+        _convertOptionsToPigeonOptions(options);
+    final SharedPreferencesAsyncApi api = _getApiForBackend(pigeonOptions);
+    return api.setInt(key, value, pigeonOptions);
   }
 
   @override
@@ -79,7 +103,10 @@ base class SharedPreferencesAsyncAndroid
     double value,
     SharedPreferencesOptions options,
   ) async {
-    return _api.setDouble(key, value, _convertOptionsToPigeonOptions(options));
+    final SharedPreferencesPigeonOptions pigeonOptions =
+        _convertOptionsToPigeonOptions(options);
+    final SharedPreferencesAsyncApi api = _getApiForBackend(pigeonOptions);
+    return api.setDouble(key, value, pigeonOptions);
   }
 
   @override
@@ -88,7 +115,10 @@ base class SharedPreferencesAsyncAndroid
     bool value,
     SharedPreferencesOptions options,
   ) async {
-    return _api.setBool(key, value, _convertOptionsToPigeonOptions(options));
+    final SharedPreferencesPigeonOptions pigeonOptions =
+        _convertOptionsToPigeonOptions(options);
+    final SharedPreferencesAsyncApi api = _getApiForBackend(pigeonOptions);
+    return api.setBool(key, value, pigeonOptions);
   }
 
   @override
@@ -97,8 +127,10 @@ base class SharedPreferencesAsyncAndroid
     List<String> value,
     SharedPreferencesOptions options,
   ) async {
-    return _api.setStringList(
-        key, value, _convertOptionsToPigeonOptions(options));
+    final SharedPreferencesPigeonOptions pigeonOptions =
+        _convertOptionsToPigeonOptions(options);
+    final SharedPreferencesAsyncApi api = _getApiForBackend(pigeonOptions);
+    return api.setStringList(key, value, pigeonOptions);
   }
 
   @override
@@ -106,8 +138,11 @@ base class SharedPreferencesAsyncAndroid
     String key,
     SharedPreferencesOptions options,
   ) async {
-    return _convertKnownExceptions<String>(() async =>
-        _api.getString(key, _convertOptionsToPigeonOptions(options)));
+    final SharedPreferencesPigeonOptions pigeonOptions =
+        _convertOptionsToPigeonOptions(options);
+    final SharedPreferencesAsyncApi api = _getApiForBackend(pigeonOptions);
+    return _convertKnownExceptions<String>(
+        () async => api.getString(key, pigeonOptions));
   }
 
   @override
@@ -115,8 +150,11 @@ base class SharedPreferencesAsyncAndroid
     String key,
     SharedPreferencesOptions options,
   ) async {
+    final SharedPreferencesPigeonOptions pigeonOptions =
+        _convertOptionsToPigeonOptions(options);
+    final SharedPreferencesAsyncApi api = _getApiForBackend(pigeonOptions);
     return _convertKnownExceptions<bool>(
-        () async => _api.getBool(key, _convertOptionsToPigeonOptions(options)));
+        () async => api.getBool(key, pigeonOptions));
   }
 
   @override
@@ -124,8 +162,11 @@ base class SharedPreferencesAsyncAndroid
     String key,
     SharedPreferencesOptions options,
   ) async {
-    return _convertKnownExceptions<double>(() async =>
-        _api.getDouble(key, _convertOptionsToPigeonOptions(options)));
+    final SharedPreferencesPigeonOptions pigeonOptions =
+        _convertOptionsToPigeonOptions(options);
+    final SharedPreferencesAsyncApi api = _getApiForBackend(pigeonOptions);
+    return _convertKnownExceptions<double>(
+        () async => api.getDouble(key, pigeonOptions));
   }
 
   @override
@@ -133,8 +174,11 @@ base class SharedPreferencesAsyncAndroid
     String key,
     SharedPreferencesOptions options,
   ) async {
+    final SharedPreferencesPigeonOptions pigeonOptions =
+        _convertOptionsToPigeonOptions(options);
+    final SharedPreferencesAsyncApi api = _getApiForBackend(pigeonOptions);
     return _convertKnownExceptions<int>(
-        () async => _api.getInt(key, _convertOptionsToPigeonOptions(options)));
+        () async => api.getInt(key, pigeonOptions));
   }
 
   @override
@@ -142,12 +186,13 @@ base class SharedPreferencesAsyncAndroid
     String key,
     SharedPreferencesOptions options,
   ) async {
+    final SharedPreferencesPigeonOptions pigeonOptions =
+        _convertOptionsToPigeonOptions(options);
+    final SharedPreferencesAsyncApi api = _getApiForBackend(pigeonOptions);
     // TODO(tarrinneal): Remove cast once https://github.com/flutter/flutter/issues/97848
     // is fixed. In practice, the values will never be null, and the native implementation assumes that.
     return _convertKnownExceptions<List<String>>(() async =>
-        (await _api.getStringList(key, _convertOptionsToPigeonOptions(options)))
-            ?.cast<String>()
-            .toList());
+        (await api.getStringList(key, pigeonOptions))?.cast<String>().toList());
   }
 
   Future<T?> _convertKnownExceptions<T>(Future<T?> Function() method) async {
@@ -169,9 +214,12 @@ base class SharedPreferencesAsyncAndroid
     SharedPreferencesOptions options,
   ) async {
     final PreferencesFilters filter = parameters.filter;
-    return _api.clear(
+    final SharedPreferencesPigeonOptions pigeonOptions =
+        _convertOptionsToPigeonOptions(options);
+    final SharedPreferencesAsyncApi api = _getApiForBackend(pigeonOptions);
+    return api.clear(
       filter.allowList?.toList(),
-      _convertOptionsToPigeonOptions(options),
+      pigeonOptions,
     );
   }
 
@@ -181,16 +229,53 @@ base class SharedPreferencesAsyncAndroid
     SharedPreferencesOptions options,
   ) async {
     final PreferencesFilters filter = parameters.filter;
-    final Map<String?, Object?> data = await _api.getAll(
+    final SharedPreferencesPigeonOptions pigeonOptions =
+        _convertOptionsToPigeonOptions(options);
+    final SharedPreferencesAsyncApi api = _getApiForBackend(pigeonOptions);
+    final Map<String?, Object?> data = await api.getAll(
       filter.allowList?.toList(),
-      _convertOptionsToPigeonOptions(options),
+      pigeonOptions,
     );
     return data.cast<String, Object>();
   }
 }
 
+/// Used to identify which Android library should be used on the backend of this call.
+enum SharedPreferencesAndroidBackendLibrary {
+  /// Represents the newer DataStore Preferences library.
+  DataStore,
+
+  /// Represents the older SharedPreferences library.
+  SharedPreferences,
+}
+
 /// Options for the Android specific SharedPreferences plugin.
 class SharedPreferencesAsyncAndroidOptions extends SharedPreferencesOptions {
   /// Constructor for SharedPreferencesAsyncAndroidOptions.
-  const SharedPreferencesAsyncAndroidOptions();
+  const SharedPreferencesAsyncAndroidOptions({
+    this.backend = SharedPreferencesAndroidBackendLibrary.DataStore,
+    this.originalSharedPreferencesOptions,
+  });
+
+  /// Which backend should be used for this method call.
+  final SharedPreferencesAndroidBackendLibrary backend;
+
+  /// These options define how the `SharedPreferences` backend should behave.
+  ///
+  /// Any options in this field will be ignored unless the backend that is selected
+  /// is `SharedPreferences`.
+  final AndroidSharedPreferencesStoreOptions? originalSharedPreferencesOptions;
+}
+
+/// Options necessary for defining the use of the original `SharedPreferences`
+/// library.
+///
+/// These options are only ever used with the original `SharedPreferences` and
+/// have no purpose when using the default DataStore Preferences.
+class AndroidSharedPreferencesStoreOptions {
+  /// Constructor for AndroidSharedPreferencesStoreOptions.
+  const AndroidSharedPreferencesStoreOptions({this.fileName});
+
+  /// The name of the file in which the preferences are stored.
+  final String? fileName;
 }