+++ packages/rfw/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Updates minimum supported SDK version to Flutter 3.22/Dart 3.4.
+
 ## 1.0.30
 
 * Adds `missing_code_block_language_in_doc_comment` lint.